make 2 files instead of one
